Remove StartIngressRequest (#179)
This removes the StartIngressRequest messages and RPC handers, as ingress will start an ingress on RTMP connection. A similar message may be needed again when we support other protocols such as SRT, but it is hard to tell exactly if the message as it exists today is the one that will be needed for this support.