`SubscriberAllowPause` as a connection option. (#346)
* `SubscriberAllowPause` as a connection option.

This allows controlling stream allocator allow pause or not decision
through a connection option.

NOTE: Thinking ahead a bit, would we allow pause or not at a track
level? We are going to do subscription priority. Maybe, we can also
allow pause or not at track level by adding it to `UpdateTrackSettings`.
Wondering if things are going to be confusing if we do that also.
I think the hierarchy would be (order of increasing precedence)
- from config
- connection option added in this PR will take effect if set
- track level allow pause config if we decide to do this (have to
  make sure that we can determine that the value was explicitly set in
  the protocol message vs using default value in protocol message).

* psrpc updated compile

* generated protobuf

* update bootstrap psrpc

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>