Rename participant_identity in Transcription
There are two fields named participant_identity in context of Transcription. It's not clear which one corresponds to the transcriber agent, and which one is for the participant which got its speech transcribed.

This change renames the Transcription field to transcribed_participant_identity so to clarify that it's used for the transcribed participant.