Use mapped SDP answer when using single peer connection mode. (#1255)
* Use mapped SDP answer when using single peer connection mode.

In single peer connection mode, the client does not get the
TR_ ids for tracks. That is because the transceiver is created with a
receiver when the offer is generated and a random id assigned. Once the
id is assigned, it is not updated. So, when client gets `onTrack`, the
id of the associated MediaStreamTrack is not the TR_ variety.

So, send the mapping of mid -> trackID in signalling when sending
answer. And used that on client side to find LK track ID for given mid.

Side note: Probably worth doing it for dual peer connection also when
subscriber sends offer so that we do not have to deduce it on client
side using some rules.

* generated protobuf

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>